requirements.txt

streamlit
pytube
openai-whisper
moviepy
torch
transformers
